promotion

+++ voyageone-service2/voyageone-service2-impl/src/main/java/com/voyageone/service/impl/cms/jumei/CmsBtJmPromotionProductService.java
@@ -70,6 +70,11 @@ public class CmsBtJmPromotionProductService {
     }
 
     public int delete(int id) {
+      CmsBtJmPromotionProductModel modelProduct=dao.select(id);
+        if(modelProduct.getSynchStatus()==2)
+        {
+            return 0;
+        }
         return dao.delete(id);
     }
